Move kits bits to own file
